refactor: changed u8 to usize

+++ ritm_core/src/parser.rs
@@ -78,7 +78,7 @@ pub fn parse_turing_machine(turing_mach: String) -> Result<TuringMachineGraph, T
                 if let None = turing_machine 
                 {
                     // With the collected number of ribbons
-                    turing_machine = Some(TuringMachineGraph::new(transitions.get(0).expect("At least one rule should be given in a transition").get_number_of_affected_ribbons() as u8).unwrap());    
+                    turing_machine = Some(TuringMachineGraph::new(transitions.get(0).expect("At least one rule should be given in a transition").get_number_of_affected_ribbons()).unwrap());    
                 }
                 // If the MT existed
                 if let Some(mt) = &mut turing_machine
